Move `Int`'s negation operations to `int::neg`
